+++ package.json
@@ -18,28 +18,19 @@
     "tailwindcss": "^3.3.3",
     "vite": "^3.2.3"
   },
-  "peerDependencies": {
-    "react": "^17.0.0 || ^18.2.0",
-    "react-dom": "^17.0.0 || ^18.2.0"
-  },
   "dependencies": {
     "@auth0/auth0-react": "^2.2.0",
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
-    "@fontsource/roboto": "^5.0.5",
-    "@mui/icons-material": "^5.14.3",
-    "@mui/material": "^5.14.3",
-    "@mui/styled-engine-sc": "^5.12.0",
-    "@mui/x-data-grid": "^6.10.2",
+    "@mui/material": "^5.14.2",
     "axios": "^1.4.0",
-    "formik": "^2.4.3",
     "gh-pages": "^5.0.0",
     "js-cookie": "^3.0.5",
     "primeflex": "^3.3.1",
     "primeicons": "^6.0.1",
     "primereact": "^9.6.0",
-    "react": "^17.0.0 || ^18.2.0",
-    "react-dom": "^17.0.0 || ^18.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.10.1",
     "react-redux": "^8.1.1",
     "react-router-dom": "^6.14.2",
@@ -47,7 +38,6 @@
     "redux": "^4.2.1",
     "redux-thunk": "^2.4.2",
     "sass": "^1.63.6",
-    "styled-components": "^5.3.11",
     "tailwindscss": "^0.3.0"
   },
   "homepage": "https://PF-Ide-Pintureria.github.io/pinturas-app-front/"
