+++ src/redux/action-type.js
@@ -5,14 +5,12 @@ export const POST_PRODUCT = "POST_PRODUCT";
 export const PUT_PRODUCT = "PUT_PRODUCT";
 export const GET_PRODUCT_BY_NAME = "GET_PRODUCT_BY_NAME";
 export const GET_PRODUCT_BY_ID = "GET_PRODUCT_BY_ID";
-export const GET_ALL_PRODUCTS_NO_FILTER = "GET_ALL_PRODUCTS_NO_FILTER";
 
 //CATEGORIES
 export const GET_ALL_CATEGORIES = "GET_ALL_CATEGORIES";
 
 //USER
 export const GET_ALL_USERS = "GET_ALL_USERS";
-export const GET_USER_BY_ID = "GET_USER_BY_ID";
 export const POST_LOGIN_USER = "POST_LOGIN_USER";
 export const POST_REGISTER_USER = "POST_REGISTER_USER";
 export const PUT_USER = "PUT_USER";
