+++ src/redux/actions/getAllUsers.js
@@ -2,18 +2,16 @@ import { GET_ALL_USERS, BASE_URL } from "../action-type";
 import axios from "axios";
 
 export const getAllUsers = () => {
-    return async (dispatch) => {
-        try {
+  return async (dispatch) => {
+    try {
 
-            const rawResponse = await axios.get(`${BASE_URL}users/`);
-            const responseData = rawResponse?.data;
-            const response = responseData?.users;
+      const response = await axios.get(`${BASE_URL}users/`).data.users;
 
-            return dispatch({ type: GET_ALL_USERS, payload: response });
+      return dispatch({ type: GET_ALL_USERS, payload: response });
 
-        } catch (error) {
-            console.error(error);
+      } catch (error) {
+        console.error(error);
 
-        }
+      }
     };
 };
